+++ client-roomfinder/src/components/Form/index.tsx
@@ -1,0 +1,12 @@
+import {FormularioProvider} from './FormularioContext';
+import Wizar from './Wizar';
+
+export default function Form(){
+    return(
+        <>
+            <FormularioProvider>
+                <Wizar/>
+            </FormularioProvider>
+        </>
+    );
+}
