+++ client-roomfinder/src/components/GeneralComponents/Button/index.tsx
@@ -1,0 +1,15 @@
+export default function Button({contenido, onClick, className}:
+    {
+        contenido:string,
+        onClick:any,
+        className:string
+    }){
+    return(
+        <button
+            onClick={onClick}
+            className={className}
+        >
+            {contenido}
+        </button>
+    );
+}
